% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_uniprot_data}
\alias{get_uniprot_data}
\title{Retrieve data from UniProt using a query}
\usage{
get_uniprot_data(
  query = NULL,
  columns = c("id", "genes", "organism", "reviewed")
)
}
\arguments{
\item{query}{A list or character containing the query for UniProt.}

\item{columns}{A vector of column names to be returned in the data frame.}
}
\value{
A data frame containing the retrieved data.
}
\description{
Retrieves data from UniProt using a query and returns the result as a data frame.
}
\examples{
query <- list(list("Aceclofenac", organism_id = "9606", reviewed = "true"))
columns <- c("accession", "gene_names", "organism_name", "reviewed")
get_uniprot_data(query[[1]], columns)
}
