% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{uniprot_drug_data}
\alias{uniprot_drug_data}
\title{Retrieve gene names and organism information of drugs from UniProt}
\usage{
uniprot_drug_data(drug)
}
\arguments{
\item{drug}{A character vector containing the names of drugs.}
}
\value{
A data frame containing the accession number, gene names, organism name, review status and the drug name.
}
\description{
Retrieves gene names and organism information of drugs from UniProt and returns the result as a data frame.
}
\examples{
drug <- c("Aceclofenac", "Benzthiazide")
uniprot_drug_data(drug)
}
